requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
asarPy==1.0.1
attrs==25.1.0
beautifulsoup4==4.13.3
bs4==0.0.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
dask==2025.1.0
dask-expr==2.0.0
dnspython==2.7.0
frozenlist==1.5.0
fsspec==2025.2.0
greenlet==3.1.1
idna==3.10
locket==1.0.0
lxml==5.3.1
multidict==6.1.0
numpy==2.2.2
packaging==24.2
pandas==2.2.3
parfive==2.1.0
partd==1.4.2
propcache==0.2.1
pyarrow==19.0.0
pymongo==4.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
soupsieve==2.6
toolz==1.0.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wget==3.2
yarl==1.18.3
